+++ src/components/gdpr/text-components/GdprText.vue
